# frontend/src/lib/helper/typegenerator/api-docs.yaml
openapi: 3.0.1
info:
  title: Tutorial Management API
  description: This API exposes endpoints to be implemented in the frontend
  version: "3.0"
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearer-key: []
...
paths:
  /add-to-dashboard:
    put:
      tags:
        - dash-board-controller
      summary: Add sensor station to dashboard
      operationId: addPlantToDashboard_1
      parameters:
        - name: person
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Person"
        - name: sensorStationId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "500":
          description: Internal Server Error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: Successfully added sensor station to dashboard
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/MessageResponse"
    post:
      tags:
        - dash-board-controller
      summary: Add sensor station to dashboard
      operationId: addPlantToDashboard
      parameters:
        - name: person
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Person"
        - name: sensorStationId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "500":
          description: Internal Server Error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: Successfully added sensor station to dashboard
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /remove-from-dashboard:
    post:
      tags:
        - dash-board-controller
      summary: Remove sensor station from dashboard
      operationId: removeFromDashboard
      parameters:
        - name: person
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Person"
        - name: sensorStationId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "500":
          description: Internal Server Error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: Successfully removed sensor station from dashboard
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      tags:
        - dash-board-controller
      summary: Remove sensor station from dashboard
      operationId: removeFromDashboard_1
      parameters:
        - name: person
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Person"
        - name: sensorStationId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "500":
          description: Internal Server Error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: Successfully removed sensor station from dashboard
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /get-dashboard:
    get:
      tags:
        - dash-board-controller
      summary: Get dashboard data
      operationId: getDashboard
      parameters:
        - name: person
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Person"
      responses:
        "500":
          description: Internal Server Error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: Successfully retrieved dashboard data
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DashBoardDataResponse"
...
components:
  schemas:
    GrantedAuthority:
      type: object
      properties:
        authority:
          type: string
    Person:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          writeOnly: true
        token:
          type: string
          format: uuid
          readOnly: true
        permissions:
          uniqueItems: true
          type: array
          items:
            $ref: "#/components/schemas/GrantedAuthority"
        email:
          type: string
        sensorStationPersonReferences:
          type: array
          items:
            $ref: "#/components/schemas/SensorStationPersonReference"
        personId:
          type: string
          format: uuid
    SensorStationPersonReference:
      type: object
      properties:
        id:
          type: string
          format: uuid
        inDashboard:
          type: boolean
        assigned:
          type: boolean
        deleted:
          type: boolean
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      description: ""
...
    DashBoardDataResponse:
      type: object
      properties:
        sensorStations:
          type: array
          items:
            $ref: "#/components/schemas/DashboardSensorStation"
      description: ""
    DashboardSensorStation:
      type: object
      properties:
        name:
          type: string
        roomName:
          type: string
        sensorStationId:
          type: string
          format: uuid
        pictureIds:
          type: array
          items:
            type: string
            format: uuid
        connected:
          type: boolean
        unlocked:
          type: boolean
        deleted:
          type: boolean
...
  securitySchemes:
    bearer-key:
      type: http
      scheme: bearer
      bearerFormat: JWT